Restoring incompatible backbone functionality
Out-of-the-box, the deferred offered from Model.fetch will have the raw ajax response as the argument passed to the callback. Instead you were passing the model, or collection itself. Although I realize its more convenient, and on a certain level, makes more sense, and this might not even be an undocumented backbone "featured", I am restoring this functionality so it works as expected.